requirements.txt
# WhatsApp Number Validation Tool - Dependencies

# Data processing
pandas>=1.3.0
openpyxl>=3.0.0

# Phone number validation
phonenumbers>=8.12.0

# API requests
requests>=2.25.0

# Configuration
python-dotenv>=0.19.0
PyYAML>=5.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0